# small blooms optimization