this pr prevents blooms filters from rebuilding when it's not necessary

importing 475k blocks from file on master:

```
200k: 1 minute 47 seconds
475k: 4 minutes 27 seconds
```

importing 475k blocks on this branch:

```
200k: 1 minute 34 seconds (0.87%)
475k: 4 minutes 27 seconds (0.93%)
```