@debris 2016-08-24
this could be further optimized by ensuring that we always store bloom filters for last 256 blocks in memory. it's gonna take 512 to 1024 kilobytes